MAINT: move CP decomposition to tensor plugin
