+++ .github/workflows/build_docs.yml
@@ -118,9 +118,10 @@ jobs:
           start=$SECONDS
           uv pip install --no-deps -e plugins/spectrochempy-nmr
           uv pip install --no-deps -e plugins/spectrochempy-iris
+          uv pip install --no-deps -e plugins/spectrochempy-tensor
           uv pip install --no-deps -e plugins/spectrochempy-hypercomplex
           # Install plugin runtime dependencies that --no-deps skips
-          uv pip install osqp scipy numpy-quaternion
+          uv pip install osqp scipy numpy-quaternion tensorly
           duration=$((SECONDS - start))
           echo "plugin installation: ${duration}s" | tee -a "$CI_TIMING_FILE"
