+++ .github/workflows/scripts/plugin_version_status.py
@@ -17,6 +17,7 @@ from pathlib import Path
 OFFICIAL_PLUGINS = (
     "spectrochempy-nmr",
     "spectrochempy-iris",
+    "spectrochempy-tensor",
     "spectrochempy-hypercomplex",
     "spectrochempy-carroucell",
 )
