+++ docs/sources/devguide/plugins/roadmap.rst
@@ -13,6 +13,7 @@ Completed decoupling
 
 * NMR unit contexts live in the NMR plugin.
 * Hypercomplex/quaternion support lives in ``spectrochempy-hypercomplex``.
+* TensorLy-backed tensor decompositions live in ``spectrochempy-tensor``.
 * 2D NMR FFT encodings are delegated through ``fft.encoding``.
 * Directory-to-file resolution for plugin formats is delegated through
   ``importer.resolve_directory_target``.
