+++ docs/sources/reference/index.rst
@@ -680,16 +680,6 @@ SIMPLe to use Interactive Self-modeling Mixture Analysis
 
     SIMPLISMA
 
-
-CP (CANDECOMP/PARAFAC) decomposition
-====================================
-
-.. autosummary::
-    :nosignatures:
-    :toctree: generated/
-
-    CP
-
 Phase Sensitive Detection
 =========================
