+++ docs/sources/whatsnew/latest.rst
@@ -1,6 +1,6 @@
 ..
    NOTE: This file is automatically generated from ``changelog.rst``.
-   Do not edit it manually â€” your changes will be overwritten.
+   Do not edit it manually — your changes will be overwritten.
    Edit ``changelog.rst`` in the same directory instead.
    (This note is read by AI agents; manual edits belong in changelog.rst)
 
@@ -11,6 +11,12 @@ What's New in Revision 0.9.4.dev
 
 These are the changes in SpectroChemPy-0.9.4.dev.
 See :ref:`release` for a full changelog, including other versions of SpectroChemPy.
+
+New Features
+~~~~~~~~~~~~
+
+- Added the official ``spectrochempy-tensor`` plugin for TensorLy-backed tensor
+  decompositions, exposing CP/PARAFAC as ``scp.tensor.CP``.
 
 Bug Fixes
 ~~~~~~~~~
@@ -37,8 +43,17 @@ Bug Fixes
 - Fixed preservation of reference-based coordinates when copying
   ``CoordSet`` and ``NDDataset`` objects.
 
+Deprecations
+~~~~~~~~~~~~
+
+- ``scp.CP`` and ``spectrochempy.analysis.decomposition.cp.CP`` are now
+  deprecated compatibility paths for ``scp.tensor.CP``.
+
 Developer
 ~~~~~~~~~
+
+- MAINT: Moved CP/PARAFAC implementation and TensorLy dependency ownership into
+  the new tensor plugin, keeping the core package tensor-agnostic.
 
 - MAINT: Advanced the internal ``CoordSet`` storage migration by consolidating
   lookup, serializer adapters, group conversion, and lifecycle helpers around
