+++ environments/environment_tensor.yml
@@ -15,14 +15,32 @@
 #
 # ======================================================================================
 
-# CORE dependencies
-# -----------------
--r requirements.txt
+channel_priority: strict
+channels:
+    - conda-forge
+    - spectrocat
+    - defaults
 
-# PLUGINS dependencies
-# --------------------
-spectrochempy[nmr]
-spectrochempy[iris]
-spectrochempy[tensor]
-spectrochempy-hypercomplex>=0.1,<0.2
-spectrochempy-carroucell>=0.1,<0.2
+dependencies:
+
+    # CORE dependencies
+    # -----------------
+    - colorama
+    - dill
+    - ipython
+    - jinja2
+    - lazy_loader
+    - matplotlib-base
+    - numpy
+    - pluggy
+    - pint
+    - pyyaml
+    - requests
+    - scikit-learn
+    - scipy
+    - tzlocal
+
+    # TENSOR dependencies
+    # -------------------
+
+    - spectrochempy-tensor>=0.1,<0.2
