+++ plugins/spectrochempy-tensor/pyproject.toml
@@ -1,0 +1,23 @@
+[build-system]
+requires = ["setuptools>=64"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "spectrochempy-tensor"
+version = "0.1.0"
+description = "Tensor decompositions plugin for SpectroChemPy"
+requires-python = ">=3.11"
+readme = "README.md"
+dependencies = [
+    "spectrochempy>=0.9,<0.10",
+    "tensorly",
+]
+
+[project.entry-points."spectrochempy.plugins"]
+tensor = "spectrochempy_tensor:TensorPlugin"
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
