+++ plugins/spectrochempy-tensor/recipe.yaml
@@ -1,0 +1,44 @@
+context:
+  name: spectrochempy-tensor
+  version: "0.1.0"
+
+package:
+  name: "${{ name }}"
+  version: "${{ version }}"
+
+source:
+  path: ../../plugins/spectrochempy-tensor
+
+build:
+  number: 0
+  noarch: python
+  script: python -m pip install . --no-deps --ignore-installed -vv
+
+requirements:
+  host:
+    - python >=3.11
+    - pip
+    - setuptools >=64
+  run:
+    - python >=3.11
+    - spectrochempy >=0.9,<0.10
+    - tensorly
+
+tests:
+  - script:
+      - python -c "import importlib.util; assert importlib.util.find_spec('spectrochempy_tensor') is not None"
+
+about:
+  summary: "Tensor decompositions plugin for SpectroChemPy"
+  description: |
+    Official SpectroChemPy plugin providing TensorLy-backed tensor
+    decomposition classes such as CP/PARAFAC, with room for future tensor
+    methods such as Tucker and TensorTrain.
+  homepage: https://www.spectrochempy.fr
+  license: LicenseRef-CeCILL-B
+  repository: https://github.com/spectrochempy/spectrochempy
+
+extra:
+  recipe-maintainers:
+    - fernandezc
+    - atravert
