+++ plugins/spectrochempy-tensor/src/spectrochempy_tensor/adapters/__init__.py
@@ -1,0 +1,3 @@
+"""Adapter layer for tensor-plugin integrations."""
+
+__all__: list[str] = []
