+++ plugins/spectrochempy-tensor/src/spectrochempy_tensor/decompositions/__init__.py
@@ -1,0 +1,5 @@
+"""TensorLy-backed tensor decomposition classes."""
+
+from spectrochempy_tensor.decompositions.cp import CP
+
+__all__ = ["CP"]
